Add support for running against py3 when run as a flake8 plugin
The plugin will first attempt to use pyqver2 as the backend to
perform the version checks. If import of this fails for any reason
(in my case because I was running using python 3 which does not
include the "compiler" module), attempt to use the pyqver3
check_versions function.

This also updates the setup.py of this repos version of pyqver to
include myself as an author and point the source back to this repo.